// src/app/pages/program-administration/models/user-info.model.ts
import { PgmPerm } from "./pgm-perm.model";

export interface UserInfo {
  userId: number;
  userName: string;
  userPwd?: string;
  fName?: string;
  lName?: string;
  displayName: string;
  secQues1?: string;
  secAns1?: string;
  secQues2?: string;
  secAns2?: string;
  addLine1?: string;
  addLine2?: string;
  city?: string;
  stateCode?: string;
  zip1?: string;
  zip2?: string;
  userPhone?: string;
  userPhoneExt?: string;
  userFax?: string;
  userEmail?: string;
  userRole: string;
  userDsg?: string;
  offCd?: string;
  parentAgy?: number;
  menuStyle: string;
  pswdResetReqd: string;
  pswdResetDt?: string;
  extRefId?: string;
  createDt?: Date;
  createId?: number;
  lastUpdDt?: Date;
  lastUpdId?: number;
  recStat: string;
  statusChangeDate?: Date;
  statusChangedBy?: number;
  menuFlag?: string;
  foodServiceDirector?: string;
  psaContractUser?: string;
  entityInformationFlag?: string;
  ctePermissions?: string;
  clnaPermissions?: string;
  miLoginAccount?: string;
  fullName?: string;
  consolidatedRvwPermCd: string;
  permCd: "";

  // statusChangedByNavigation: UserInfo;
  // inverseStatusChangedByNavigation: UserInfo[];
  // pgmPerms: PgmPerm[];
}